packet: create bmp package
move bmp stuff from bgp to bmp package.
