+++ packet/bmp/bmp_test.go
@@ -13,9 +13,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package bgp
+package bmp
 
 import (
+	"github.com/osrg/gobgp/packet/bgp"
 	"github.com/stretchr/testify/assert"
 	"reflect"
 	"testing"
@@ -46,7 +47,7 @@ func Test_Initiation(t *testing.T) {
 }
 
 func Test_PeerUpNotification(t *testing.T) {
-	m := open()
+	m := bgp.NewTestBGPOpenMessage()
 	p0 := NewBMPPeerHeader(0, false, 1000, "10.0.0.1", 70000, "10.0.0.2", 1)
 	verify(t, NewBMPPeerUpNotification(*p0, "10.0.0.3", 10, 100, m, m))
 	p1 := NewBMPPeerHeader(0, false, 1000, "fe80::6e40:8ff:feab:2c2a", 70000, "10.0.0.2", 1)
@@ -56,12 +57,12 @@ func Test_PeerUpNotification(t *testing.T) {
 func Test_PeerDownNotification(t *testing.T) {
 	p0 := NewBMPPeerHeader(0, false, 1000, "10.0.0.1", 70000, "10.0.0.2", 1)
 	verify(t, NewBMPPeerDownNotification(*p0, BMP_PEER_DOWN_REASON_UNKNOWN, nil, []byte{0x3, 0xb}))
-	m := NewBGPNotificationMessage(1, 2, nil)
+	m := bgp.NewBGPNotificationMessage(1, 2, nil)
 	verify(t, NewBMPPeerDownNotification(*p0, BMP_PEER_DOWN_REASON_LOCAL_BGP_NOTIFICATION, m, nil))
 }
 
 func Test_RouteMonitoring(t *testing.T) {
-	m := update()
+	m := bgp.NewTestBGPUpdateMessage()
 	p0 := NewBMPPeerHeader(0, false, 1000, "fe80::6e40:8ff:feab:2c2a", 70000, "10.0.0.2", 1)
 	verify(t, NewBMPRouteMonitoring(*p0, m))
 }
